// orgbac-practitioner-application/fhir.r4.sdk/types/hl7-fhir-r4-core/PractitionerRole.ts
// WARNING: This file is autogenerated by FHIR Schema Codegen.
// https://github.com/fhir-schema/fhir-schema-codegen
// Any manual changes made to this file may be overwritten.

import { BackboneElement } from './BackboneElement';
import { CodeableConcept } from './CodeableConcept';
import { ContactPoint } from './ContactPoint';
import { DomainResource } from './DomainResource';
import { Identifier } from './Identifier';
import { Period } from './Period';
import { Reference } from './Reference';


export interface PractitionerRoleAvailableTime extends BackboneElement {
    allDay?: boolean;
    availableEndTime?: string;
    availableStartTime?: string;
    daysOfWeek?: 'mon' | 'tue' | 'wed' | 'thu' | 'fri' | 'sat' | 'sun'[];
}

export interface PractitionerRoleNotAvailable extends BackboneElement {
    description: string;
    during?: Period;
}

export interface PractitionerRole extends DomainResource {
    active?: boolean;
    _active?: Element;
    availabilityExceptions?: string;
    _availabilityExceptions?: Element;
    availableTime?: PractitionerRoleAvailableTime[];
    code?: CodeableConcept[];
    endpoint?: Reference<'Endpoint'>[];
    healthcareService?: Reference<'HealthcareService'>[];
    identifier?: Identifier[];
    location?: Reference<'Location'>[];
    notAvailable?: PractitionerRoleNotAvailable[];
    organization?: Reference<'Organization'>;
    period?: Period;
    practitioner?: Reference<'Practitioner'>;
    specialty?: CodeableConcept[];
    telecom?: ContactPoint[];
}